% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal_window.R
\name{modal_window}
\alias{modal_window}
\title{Modal}
\usage{
modal_window(
  item_id,
  windowCondition = TRUE,
  input_id_cancel = "cancel_modal",
  input_id_add = "added_note_info",
  obj = NULL,
  ...
)
}
\description{
Create modal
}
