% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slider_.R
\name{slider_Input}
\alias{slider_Input}
\title{Slider}
\usage{
slider_Input(
  inputId,
  default = 2010,
  secondaryColor = "#ffffff",
  mainColor = "#e05151",
  disabled = FALSE
)
}
\description{
Material Slider
}
