% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggl_btn_group_.R
\name{toggl_btn_group_Input}
\alias{toggl_btn_group_Input}
\title{Toggl Group Button}
\usage{
toggl_btn_group_Input(
  inputId,
  value,
  fontColor = "#ffffff",
  mainColor = "#e05151"
)
}
\description{
Toggl Group button
}
