% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggl_btn_.R
\name{toggl_btn_Input}
\alias{toggl_btn_Input}
\title{Toggl Button}
\usage{
toggl_btn_Input(
  inputId,
  label,
  default = TRUE,
  labelPlacement = c("end", "start", "top", "bottom"),
  fontColor = "#ffffff",
  mainColor = "#e05151"
)
}
\description{
Create toggle button
}
